FIxed the umoci issue
Fixed the umoci issue and modified the files to support the file structure
