+++ unit_tests/L1_testing/test_files/test_Utils_ut.py
@@ -23,11 +23,11 @@ import unittest
 
 sys.path.insert(0, os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
 from get_L1_test_results import add_test_results
-from bundlegen.core.stb_platform import STBPlatform
+from bundlegen.core.utils import Utils
 from loguru import logger
 
 #This class will test the functionality of API's in stbplatform.py file.
-class TestStbPlatform(unittest.TestCase):
+class TestUtils(unittest.TestCase):
     def setUp(self):
          logger.debug("Setup")
          add_test_results.add_tests(self)
@@ -61,37 +61,48 @@ class TestStbPlatform(unittest.TestCase):
     def tearDownClass(self):
         add_test_results.end_results(self)
 
-    def test_platform_config_schema(self):
-        ''''this test is to check is jsonschema of tempete platform schema is proper.
-        one more changes is been added chdir because test been running in folder /BundleGen/unit_tests/L1_testing
-        to validate schema changing the directory to /BundleGen
-        '''
-        logger.debug("==> Validating the templete config schema%s \n "%(os.getcwd()))
-        os.chdir('../../')
-        search_path = os.path.abspath(os.path.join( os.getcwd(),'unit_tests','L1_testing','test_data_files'))
-        logger.debug(f"==> creating the search config file path : {search_path}")
-        validate = STBPlatform("rpi3_reference_vc4_dunfell",search_path)
-        actual = validate.validate_platform_config()
-        expected = True
-        os.chdir('unit_tests/L1_testing')
-        self.assertEqual(actual, expected)
-        logger.debug("-->Test was Successfully verified")
+    def test_to_create_ipk_file(self):
+        logger.debug("-->checking to create the ipk file ")
+        source = "./test_data_files/dac-image-wayland-egl-test-bundle"
+        final_app_metadata = {
+            "id": "com.rdk.wayland-egl-test",
+            "type": "application/vnd.rdk-app.dac.native",
+            "version": "1.0.0",
+            "description": "Simple wayland egl demo, showing green rectangle",
+            "priority": "optional",
+            "graphics": True
+        }
+        platform_cfg = {
+            "platformName": "rpi3_reference",
+            "os": "linux",
+            "arch": {
+                "arch": "arm",
+                "variant": "v7"
+            }
+        }
+        Utils.create_control_file(platform_cfg, final_app_metadata)
+        expected = Utils.create_ipk(source, source)
+        self.assertEqual(True, expected)
 
+    def test_failed_to_create_tar_source_file(self):
+        logger.debug("-->checking to create the ipk file ")
+        source = "./dac-image-wayland-egl-test-bundle"
+        DATA_NAME = "data.tar.gz"
+        expected = Utils.create_tgz(source, DATA_NAME)
+        self.assertEqual(False, expected)
 
-    def test_wrong_platform_config_schema(self):
-        ''''this test is to validate if jsonschema of tempete plateform is proper.
-        here i have used vagrant file for testing because in rdk feild "supportedFeatures" is required that
-        feild was not there so used.
-        '''
-        logger.debug("==> Validating the wrong templete config schema")
-        os.chdir('../../')
-        search_path = os.path.abspath(os.path.join( os.getcwd(),'unit_tests','L1_testing','test_data_files'))
-        validate = STBPlatform("vagrant",search_path)
-        actual = validate.validate_platform_config()
-        expected = False
-        self.assertEqual(actual, expected)
-        os.chdir('unit_tests/L1_testing')
-        logger.debug("-->Test was Successfully verified")
+    def test_failed_to_run_process_and_return_output(self):
+        logger.debug("-->checking to run the command ")
+        command = "umoci unpack --rootless --image /tmp/bundlegen/20230331-112041_ae6ca9eeab67494e9cec8b206bcef66b:latest ./BundleGen/dac-image-wayland-egl-test-bundle"
+        expected = Utils.run_process_and_return_output(command)
+        self.assertEqual((1, ''), expected)
+
+    def test_utils_add_tarinfo_case(self):
+        logger.debug("-->checking new api tarinfo of utils file ")
+        source = "./test_data_files/dac-image-wayland-egl-test-bundle"
+        DATA_NAME = "data.tar.gz"
+        expected = Utils.create_tgz(source, DATA_NAME, 1, 2, '770')
+        self.assertEqual(True, expected)
 
 
 if __name__ == "__main__":
